# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZEROHOUR REALTOR PULSE API",
    "version": "0.1.0",
    "description": "Phase 1 local MVP API contract for ZEROHOUR REALTOR PULSE. This contract supports frontend integration, simulated creator onboarding, async provisioning, mock ZIP signal ingestion, cached dashboards, signal actions, billing handoff, and basic admin operations."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development API"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login with email and password",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "agent@example.com",
                "password": "correct-horse-123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/onboarding/provisioning/{job_id}": {
      "get": {
        "tags": ["onboarding"],
        "summary": "Get async provisioning job status",
        "operationId": "getProvisioningStatus",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provisioning status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisioningStatusResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/onboarding/dre-verification": {
      "post": {
        "tags": ["onboarding"],
        "summary": "Submit DRE/license verification for async processing",
        "operationId": "submitDreVerification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DreVerificationRequest"
              },
              "example": {
                "state": "CA",
                "license_number": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DreVerificationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/zips/{zip_code}": {
      "get": {
        "tags": ["dashboard"],
        "summary": "Get cached ZIP-level signal payload",
        "operationId": "getZipSignals",
        "parameters": [
          {
            "name": "zip_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{5}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ZIP cache payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZipSignalsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/signals": {
      "get": {
        "tags": ["signals"],
        "summary": "List active signals",
        "operationId": "listSignals",
        "parameters": [
          {
            "name": "zip_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{5}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active signal list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized"
      },
      "NotFound": {
        "description": "Not found"
      },
...
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["access_token", "token_type"],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "default": "bearer"
          }
        }
      },
      "ProvisioningStatusResponse": {
        "type": "object",
        "required": ["id", "status", "requested_zips"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "processing", "ready", "failed"]
          },
          "message": {
            "type": ["string", "null"]
          },
          "requested_zips": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DreVerificationRequest": {
        "type": "object",
        "required": ["state", "license_number"],
        "properties": {
          "state": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2
          },
          "license_number": {
            "type": "string"
          }
        }
      },
      "DreVerificationResponse": {
        "type": "object",
        "required": ["verification_id", "status", "message"],
        "properties": {
          "verification_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ZipSignalsResponse": {
        "type": "object",
        "required": ["zip_code", "cache_status", "payload"],
        "properties": {
          "zip_code": {
            "type": "string"
          },
          "cache_status": {
            "type": "string",
            "enum": ["redis", "postgres_stale_ok", "degraded_empty"]
          },
          "payload": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "SignalSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "zip_code": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "confidence": {
            "type": "number"
          },
          "freshness_days": {
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "address": {
            "type": ["string", "null"]
          },
          "signal_fired_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "signal_claimed_at": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "claimed_by_user_id": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "claim_delta_minutes": {
            "type": ["integer", "null"]
          }
        },
        "additionalProperties": true
      },
...
      "SignalListResponse": {
        "type": "object",
        "required": ["signals"],
        "properties": {
          "signals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SignalSummary"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}